# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GNews API",
    "version": "1.0.0",
    "description": "A wrapper for GNews API with in-memory caching"
  },
  "paths": {
    "/articles": {
      "get": {
        "summary": "Get Top Headlines",
        "description": "Fetch top news articles using optional filters. Requires 'n' (number of articles). Results are cached for 5 minutes.",
        "parameters": [
          {
            "in": "query",
            "name": "n",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Number of articles to return",
            "example": 5
          },
          {
            "in": "query",
            "name": "category",
            "schema": { "type": "string" },
            "description": "Category of news (e.g., general, technology, sports)",
            "example": "technology"
          },
          {
            "in": "query",
            "name": "lang",
            "schema": { "type": "string" },
            "description": "Language code (e.g., en, es)",
            "example": "en"
          },
          {
            "in": "query",
            "name": "country",
            "schema": { "type": "string" },
            "description": "Country code (e.g., us, ke)",
            "example": "us"
          }
        ],
        "tags": ["Articles"],
        "responses": {
          "200": {
            "description": "A list of top headlines",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Article" }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameter 'n'"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search News Articles",
        "description": "Search for news articles by keyword. Results are cached for 5 minutes.",
        "tags": ["Articles"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": { "type": "string" },
            "description": "Search query string",
            "example": "bitcoin"
          }
        ],
        "responses": {
          "200": {
            "description": "List of search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Article" }
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "image": { "type": "string", "format": "uri" },
          "publishedAt": { "type": "string", "format": "date-time" },
          "source": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "url": { "type": "string", "format": "uri" }
            }
          }
        },
        "required": ["title", "url", "publishedAt"]
      }
    }
  }
}